# openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "additionalProperties": false,
        "description": "One account in the org’s chart of accounts.",
        "properties": {
          "code": {
            "description": "Short reference unique within the org, e.g. `1000`. Compared under the column's `utf8mb4_0900_ai_ci` collation, so it is case- and accent-insensitive: `1000a` and `1000A` are the same code. Leading and trailing whitespace is trimmed.",
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "description": "Optional free text. Send `null` to clear it.",
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "isActive": {
            "description": "Inactive accounts keep every posting they carry and cannot be selected for new ones. This is the only form of removal available to an account that has been posted to.",
            "type": "boolean"
          },
          "name": {
            "description": "Display name, e.g. `Operating bank account`.",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "normalBalance": {
            "description": "The side that increases this account. Stored, never derived from `type`, because contra accounts are real: accumulated depreciation is an `asset` whose normal balance is `credit`, as is an allowance for doubtful accounts. No constraint ties the two fields together — see the `accounts` commentary in migration 0002_ledger.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          },
          "type": {
            "description": "Which of the five statement categories the account belongs to. Determines where it appears in reports; it does not determine `normalBalance`.",
            "enum": [
              "asset",
              "liability",
              "equity",
              "revenue",
              "expense"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "type",
          "normalBalance",
          "description",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AccountList": {
        "additionalProperties": false,
        "description": "The matching accounts, ordered by `code`. Unpaginated in M1: a chart of accounts is bounded by the org’s own chart.",
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/Account"
            },
            "type": "array"
          }
        },
        "required": [
          "accounts"
        ],
        "type": "object"
      },
...
      "CreateAccountRequestInput": {
        "additionalProperties": false,
        "description": "Creates one account. Accounts are created active.",
        "properties": {
          "code": {
            "description": "Short reference unique within the org, e.g. `1000`. Compared under the column's `utf8mb4_0900_ai_ci` collation, so it is case- and accent-insensitive: `1000a` and `1000A` are the same code. Leading and trailing whitespace is trimmed.",
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "description": "Optional free text. Send `null` to clear it.",
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "description": "Display name, e.g. `Operating bank account`.",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "normalBalance": {
            "description": "The side that increases this account. Stored, never derived from `type`, because contra accounts are real: accumulated depreciation is an `asset` whose normal balance is `credit`, as is an allowance for doubtful accounts. No constraint ties the two fields together — see the `accounts` commentary in migration 0002_ledger.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          },
          "type": {
            "description": "Which of the five statement categories the account belongs to. Determines where it appears in reports; it does not determine `normalBalance`.",
            "enum": [
              "asset",
              "liability",
              "equity",
              "revenue",
              "expense"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "type",
          "normalBalance"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every non-2xx response.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable failure code. Integrators branch on this, not on the HTTP status: `conflict` and `idempotency_key_conflict` are both 409 and have different recoveries. Codes are never renamed and never change status.",
                "enum": [
                  "validation_failed",
                  "unauthenticated",
                  "permission_denied",
                  "not_found",
                  "conflict",
                  "idempotency_key_conflict",
                  "precondition_failed",
                  "internal_error"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "UpdateAccountRequestInput": {
        "additionalProperties": false,
        "description": "Partial update. An absent field is unchanged; `description: null` clears it. `type` and `normalBalance` are refused once the account has postings.",
        "properties": {
          "code": {
            "description": "Short reference unique within the org, e.g. `1000`. Compared under the column's `utf8mb4_0900_ai_ci` collation, so it is case- and accent-insensitive: `1000a` and `1000A` are the same code. Leading and trailing whitespace is trimmed.",
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "description": "Optional free text. Send `null` to clear it.",
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "description": "Display name, e.g. `Operating bank account`.",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "normalBalance": {
            "description": "The side that increases this account. Stored, never derived from `type`, because contra accounts are real: accumulated depreciation is an `asset` whose normal balance is `credit`, as is an allowance for doubtful accounts. No constraint ties the two fields together — see the `accounts` commentary in migration 0002_ledger.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          },
          "type": {
            "description": "Which of the five statement categories the account belongs to. Determines where it appears in reports; it does not determine `normalBalance`.",
            "enum": [
              "asset",
              "liability",
              "equity",
              "revenue",
              "expense"
            ],
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Open-source, AI-native, modular double-entry accounting. Every capability in the product is reachable through this API (spec §8) — the web client is a consumer of it, not a privileged path.\n\nMonetary amounts are strings of minor units (`\"1999\"` is 19.99 in a two-decimal currency), never JSON numbers. A JSON number cannot represent a cent exactly and any client that parses one has already lost the value.\n\nEvery write endpoint requires an `Idempotency-Key` header. Reusing a key with the same request replays the original outcome; reusing it with a different request is refused with `idempotency_key_conflict`.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "OpenBooks API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "operationId": "listAccounts",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "asset",
                "liability",
                "equity",
                "revenue",
                "expense"
              ],
              "type": "string"
            }
          },
          {
            "description": "Accepts `true`/`false` (and `1`/`0`, `yes`/`no`, `on`/`off`). Omitted matches active and inactive accounts alike.",
            "in": "query",
            "name": "isActive",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountList"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List the chart of accounts",
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "description": "Accounts are created active. `normalBalance` is required and independent of `type`, because contra accounts are real — accumulated depreciation is an `asset` whose normal balance is `credit`.",
        "operationId": "createAccount",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create an account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/v1/accounts/{accountId}": {
      "delete": {
        "description": "An account is configuration rather than a record of what happened, so deleting an unreferenced one restates nothing. An account with postings answers `precondition_failed` with `account_has_postings`; deactivate it instead.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Delete an account that has never been posted to",
        "tags": [
          "accounts"
        ]
      },
      "get": {
        "operationId": "getAccount",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "One account",
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "description": "`type` and `normalBalance` are refused once the account carries postings: changing either would silently restate what past reports said about entries already made.",
        "operationId": "updateAccount",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Update an account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/v1/accounts/{accountId}/deactivate": {
      "post": {
        "description": "Removes an account from circulation without removing it from the books, which is the only form of removal available to an account that has been posted to. Idempotent: an already-inactive account is returned unchanged rather than refused.",
        "operationId": "deactivateAccount",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Deactivate an account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/v1/accounts/{accountId}/reactivate": {
      "post": {
        "description": "The counterpart to deactivation, so that deactivation is not a one-way door.",
        "operationId": "reactivateAccount",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Reactivate an account",
        "tags": [
          "accounts"
        ]
      }
    },
...
  },
...
}